refactor(ui): remove old (pre-JN) onboarding page and associated components
